move fixDataPage so that it execute only  in 13.4.13+
